+++ retail/BetterBlizzPlates.lua
@@ -12,7 +12,7 @@ LSM:Register("font", "Prototype", [[Interface\Addons\BetterBlizzPlates\media\Pro
 
 local addonVersion = "1.00" --too afraid to to touch for now
 local addonUpdates = C_AddOns.GetAddOnMetadata("BetterBlizzPlates", "Version")
-local sendUpdate = false
+local sendUpdate = true
 BBP.VersionNumber = addonUpdates
 local _, playerClass
 local playerClassColor
@@ -461,6 +461,7 @@ local defaultSettings = {
     nameplateAuraRowAbove = true,
     nameplateAuraHeightGap = 4,
     nameplateAuraWidthGap = 4,
+    nameplateAuraTypeGap = 0,
     nameplateAurasYPos = 0,
     nameplateAurasXPos = 0,
     nameplateAurasPersonalXPos = 0,
@@ -1266,14 +1267,14 @@ local function SendUpdateMessage()
                 -- DEFAULT_CHAT_FRAME:AddMessage("   - Fix aura color module not working on buffs.")
                 -- DEFAULT_CHAT_FRAME:AddMessage("   - Fix class icon module causing a lua error sometimes.")
 
-                if BetterBlizzPlatesDB.useCustomCastbarBGTexture then
-                    StaticPopupDialogs["BBP_BG_COLOR_UPDATE"] = {
-                        text = "|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rPlates: \n\nNameplate castbar background color is no longer red by default on un-interruptible casts.\n\nIf you want this back you will have to go to:\n/bbp -> Castbar -> Right-click the color button next to background texture to re-enable it.",
-                        button1 = "Ok",
-                        timeout = 0,
-                        whileDead = true,
+                if BetterBlizzPlatesDB.enableNameplateAuraCustomisation and BetterBlizzPlatesDB.separateAuraBuffRow then
+                    StaticPopupDialogs["BBP_ROW_UPDATE"] = {
+                    text = "|A:gmchat-icon-blizz:16:16|a Better|cff00c0ffBlizz|rPlates: \n\nFixed an issue causing Nameplate Buffs with separate buff row enabled to move upwards on higher scale settings.\n\nThis will likely have shifted the position of them for many setups.\n\nAlso introduced a new\n\"Gap between Buffs & Debuffs\"\nslider to adjust the space to your liking.\n|cff808080Also accepts negative values if you want to have them below (rows will still stack upwards tho)|r",
+                    button1 = "Ok",
+                    timeout = 0,
+                    whileDead = true,
                     }
-                    StaticPopup_Show("BBP_BG_COLOR_UPDATE")
+                    StaticPopup_Show("BBP_ROW_UPDATE")
                 end
             end)
         else