+++ temp_era/modules/auras.lua
@@ -1858,9 +1858,10 @@ local function CustomBuffLayoutChildrenCata(container, children, isEnemyUnit)
         local horizontalOffset = 0
         local firstRowFirstAuraOffset = nil  -- Variable to store the horizontal offset of the first aura in the first row
         local nameplateAurasFriendlyCenteredAnchor = db.nameplateAurasFriendlyCenteredAnchor and not isEnemyUnit
-        local nameplateAurasEnemyCenteredAnchor = db.nameplateAurasEnemyCenteredAnchor and isEnemyUnit
+        local nameplateAurasEnemyCenteredAnchor = db.nameplateAurasEnemyCenteredAnchor and isEnemyUnit and (not db.nameplateCenterOnlyBuffs or isBuffs)
         local nameplateCenterAllRows = db.nameplateCenterAllRows and (nameplateAurasFriendlyCenteredAnchor or nameplateAurasEnemyCenteredAnchor)
         local xPos = db.nameplateAurasXPos
+        local nameplateAuraTypeGap = db.nameplateAuraTypeGap
         local compactTracker = 0
         local visibleIndex = 0
 
@@ -1915,7 +1916,7 @@ local function CustomBuffLayoutChildrenCata(container, children, isEnemyUnit)
                     maxDebuffHeight = maxRowHeight
                     verticalOffset = -currentRow * (-maxRowHeight + (currentRow > 0 and verticalSpacing or 0))
                 else
-                    verticalOffset = -currentRow * (-maxDebuffHeight + (currentRow > 0 and verticalSpacing or 0))
+                    verticalOffset = (-currentRow * (-maxDebuffHeight + (currentRow > 0 and verticalSpacing or 0))) + nameplateAuraTypeGap
                 end
 
                 local extraOffset = 0